requirements.txt

Django>=4.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
redis
    